# backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MD Office API",
    "version": "1.0.0",
    "description": "REST API for MD Office document management"
  },
  "servers": [
    { "url": "/api/v1", "description": "API v1" }
  ],
  "security": [
    { "apiKey": [] }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key (mdo_...)"
      },
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/search": {
      "get": {
        "summary": "Search across all document types",
        "operationId": "search",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "Search results" } }
      }
    },
    "/export/{type}/{id}": {
      "get": {
        "summary": "Export document",
        "operationId": "exportDocument",
        "parameters": [
          { "name": "type", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["markdown", "html", "json"], "default": "markdown" } }
        ],
        "responses": { "200": { "description": "Exported document" } }
      }
    }
  }
}